pyproject.toml
[tool.poetry]
name = "freebooter"
version = "0.1.0"
description = "freebooter downloads photos & videos from the internet and uploads it onto your social media accounts."
authors = ["Parker Wahle <regulad@regulad.xyz>"]
license = "GPLv3"
readme = "README.md"
packages = [{ include = "freebooter" }]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Artistic Software",
    "Topic :: Multimedia :: Video",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.11"
youtube-dl = "^2021.12.17"
mariadb = "^1.1.5.post3"
poetry = "^1.3.1"
google-api-python-client = "^2.70.0"
dislog = "^2.0.0"
jsonschema = "^4.17.3"
google-auth = {extras = ["reauth"], version = "^2.15.0"}
google-auth-oauthlib = "^0.8.0"
httplib2 = "^0.21.0"

[tool.poetry.scripts]
freebooter = "freebooter.__main__:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
google-api-python-client-stubs = "^1.15.0"
mypy = "^0.991"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
